+++ src/templates/TemplateGroupUtils.ts
@@ -1,10 +1,9 @@
 import type { TemplateGroupRoot } from "../api/entities/tutanota/TypeRefs.js"
 import { TemplateGroupRootTypeRef } from "../api/entities/tutanota/TypeRefs.js"
-import { showPlanUpgradeRequiredDialog } from "../misc/SubscriptionDialogs"
+import { getAvailableMatchingPlans, showPlanUpgradeRequiredDialog } from "../misc/SubscriptionDialogs"
 import { locator } from "../api/main/MainLocator"
 import { FeatureType, PlanType } from "../api/common/TutanotaConstants"
 import { isCustomizationEnabledForCustomer } from "../api/common/utils/Utils"
-import { getAvailableMatchingPlans } from "../subscription/SubscriptionUtils.js"
 
 /**
  * @return True if the group has been created.
